backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.2.1

# Database
motor==3.4.0
pymongo==4.7.2

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Face Recognition
opencv-python-headless==4.9.0.80
face-recognition==1.3.0
numpy==1.26.4
Pillow==10.3.0

# AI
google-generativeai==0.5.4

# Utilities
python-dotenv==1.0.1
requests==2.31.0
